#68 - move sarif readers to separate package
