+++ plugin/src/test/java/org/owasp/benchmarkutils/score/parsers/sarif/PrecautionReaderTest.java
@@ -12,12 +12,13 @@
  * WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR
  * PURPOSE. See the GNU General Public License for more details.
  *
- * @author Nicolas Couraud
- * @created 2023
+ * @author Eric Brown
+ * @created 2024
  */
-package org.owasp.benchmarkutils.score.parsers;
+package org.owasp.benchmarkutils.score.parsers.sarif;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -26,53 +27,35 @@ import org.owasp.benchmarkutils.score.CweNumber;
 import org.owasp.benchmarkutils.score.ResultFile;
 import org.owasp.benchmarkutils.score.TestHelper;
 import org.owasp.benchmarkutils.score.TestSuiteResults;
+import org.owasp.benchmarkutils.score.parsers.ReaderTestBase;
+import org.owasp.benchmarkutils.score.parsers.sarif.PrecautionReader;
 
-import java.io.File;
-import java.io.IOException;
-
-public class CodeQLReaderTest extends ReaderTestBase {
+class PrecautionReaderTest extends ReaderTestBase {
 
     private ResultFile resultFile;
 
     @BeforeEach
     void setUp() {
-        resultFile = TestHelper.resultFileOf("testfiles/Benchmark_CodeQL-v2.13.sarif");
+        resultFile = TestHelper.resultFileOf("testfiles/Benchmark_Precaution.sarif");
         BenchmarkScore.TESTCASENAME = "BenchmarkTest";
     }
 
     @Test
-    public void onlyCodeQLReaderTestReportsCanReadAsTrue() {
-        assertOnlyMatcherClassIs(this.resultFile, CodeQLReader.class);
+    public void onlyPrecautionReportsCanReadAsTrue() {
+        assertOnlyMatcherClassIs(this.resultFile, PrecautionReader.class);
     }
 
     @Test
     void readerHandlesGivenResultFile() throws Exception {
-        CodeQLReader reader = new CodeQLReader();
+        PrecautionReader reader = new PrecautionReader();
         TestSuiteResults result = reader.parse(resultFile);
 
         assertEquals(TestSuiteResults.ToolType.SAST, result.getToolType());
+        assertFalse(result.isCommercial());
+        assertEquals("Precaution", result.getToolName());
+        assertEquals("0.5.0", result.getToolVersion());
 
-        assertEquals("CodeQL", result.getToolName());
-
-        assertEquals(2, result.getTotalResults());
-
-        assertEquals(CweNumber.XSS, result.get(1).get(0).getCWE());
-        assertEquals(CweNumber.SQL_INJECTION, result.get(2).get(0).getCWE());
-    }
-
-    @Test
-    void readerHandlesAlternativeResultFile() throws Exception {
-        resultFile = TestHelper.resultFileOf("testfiles/Benchmark_CodeQL-v2.13.alternative.sarif");
-        CodeQLReader reader = new CodeQLReader();
-        TestSuiteResults result = reader.parse(resultFile);
-
-        assertEquals(TestSuiteResults.ToolType.SAST, result.getToolType());
-
-        assertEquals("CodeQL", result.getToolName());
-
-        assertEquals(2, result.getTotalResults());
-
-        assertEquals(CweNumber.XSS, result.get(1).get(0).getCWE());
-        assertEquals(CweNumber.SQL_INJECTION, result.get(2).get(0).getCWE());
+        assertEquals(1, result.getTotalResults());
+        assertEquals(CweNumber.WEAK_HASH_ALGO, result.get(73).get(0).getCWE());
     }
 }
