+++ plugin/src/test/java/org/owasp/benchmarkutils/score/parsers/sarif/SnykReaderTest.java
@@ -10,15 +10,15 @@
  *
  * <p>The OWASP Benchmark is distributed in the hope that it will be useful, but WITHOUT ANY
  * WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR
- * PURPOSE. See the GNU General Public License for more details.
+ * PURPOSE. See the GNU General Public License for more details
  *
- * @author Eric Brown
- * @created 2024
+ * @author Raj Barath
+ * @created 2023
  */
-package org.owasp.benchmarkutils.score.parsers;
+package org.owasp.benchmarkutils.score.parsers.sarif;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -27,33 +27,36 @@ import org.owasp.benchmarkutils.score.CweNumber;
 import org.owasp.benchmarkutils.score.ResultFile;
 import org.owasp.benchmarkutils.score.TestHelper;
 import org.owasp.benchmarkutils.score.TestSuiteResults;
+import org.owasp.benchmarkutils.score.parsers.ReaderTestBase;
 
-class PrecautionReaderTest extends ReaderTestBase {
+class SnykReaderTest extends ReaderTestBase {
 
     private ResultFile resultFile;
 
     @BeforeEach
     void setUp() {
-        resultFile = TestHelper.resultFileOf("testfiles/Benchmark_Precaution.sarif");
+        resultFile = TestHelper.resultFileOf("testfiles/Benchmark_SnykCodeCli.sarif");
         BenchmarkScore.TESTCASENAME = "BenchmarkTest";
     }
 
     @Test
-    public void onlyPrecautionReportsCanReadAsTrue() {
-        assertOnlyMatcherClassIs(this.resultFile, PrecautionReader.class);
+    void onlySnykReaderReportsCanReadAsTrue() {
+        assertOnlyMatcherClassIs(this.resultFile, SnykReader.class);
     }
 
     @Test
     void readerHandlesGivenResultFile() throws Exception {
-        PrecautionReader reader = new PrecautionReader();
+        SnykReader reader = new SnykReader();
         TestSuiteResults result = reader.parse(resultFile);
 
         assertEquals(TestSuiteResults.ToolType.SAST, result.getToolType());
-        assertFalse(result.isCommercial());
-        assertEquals("Precaution", result.getToolName());
-        assertEquals("0.5.0", result.getToolVersion());
+        assertTrue(result.isCommercial());
+        assertEquals("SnykCode", result.getToolName());
+        assertEquals("1.0.0", result.getToolVersion());
 
-        assertEquals(1, result.getTotalResults());
-        assertEquals(CweNumber.WEAK_HASH_ALGO, result.get(73).get(0).getCWE());
+        assertEquals(2, result.getTotalResults());
+
+        assertEquals(CweNumber.INSECURE_COOKIE, result.get(1).get(0).getCWE());
+        assertEquals(CweNumber.XPATH_INJECTION, result.get(2).get(0).getCWE());
     }
 }
